refactor: split context into dedicated folder + rename paths + rollup config
