+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './ui/index'
-export * from './hooks/index'
+export * from './hook/index'
+export * from './context/index'
 export type { DeepReadonly } from './superTypes'
