+++ src/ui/index.ts
@@ -2,4 +2,3 @@ export * from './atoms/button/Button'
 export * from './atoms/canvas/Canvas'
 export * from './atoms/theme/ThemeProvider'
 export * from './atoms/theme/ThemeSwitcher'
-export type { Theme, ThemeContextType } from './atoms/theme/themeContext'
